Refactored operation outcomes in aggregation
