+++ dpc-aggregation/src/main/java/gov/cms/dpc/aggregation/engine/JobBatchProcessor.java
@@ -5,6 +5,7 @@ import com.codahale.metrics.Meter;
 import com.codahale.metrics.MetricRegistry;
 import com.google.common.net.HttpHeaders;
 import gov.cms.dpc.aggregation.service.*;
+import gov.cms.dpc.aggregation.util.AggregationUtils;
 import gov.cms.dpc.bluebutton.client.BlueButtonClient;
 import gov.cms.dpc.common.Constants;
 import gov.cms.dpc.common.MDCConstants;
@@ -15,8 +16,6 @@ import gov.cms.dpc.queue.models.JobQueueBatchFile;
 import io.reactivex.Flowable;
 import org.apache.commons.lang3.tuple.Pair;
 import org.hl7.fhir.dstu3.model.*;
-import org.hl7.fhir.dstu3.model.OperationOutcome.IssueSeverity;
-import org.hl7.fhir.dstu3.model.OperationOutcome.IssueType;
 import org.reactivestreams.Publisher;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -62,23 +61,26 @@ public class JobBatchProcessor {
      * @return A list of batch files {@link JobQueueBatchFile}
      */
     public List<JobQueueBatchFile> processJobBatchPartial(UUID aggregatorID, IJobQueue queue, JobQueueBatch job, String patientID) {
+        OutcomeReason failReason = null;
         final Pair<Optional<List<ConsentResult>>,Optional<OperationOutcome>> consentResult = getConsent(patientID);
 
         Flowable<Resource> flowable;
         if(consentResult.getRight().isPresent()){
             flowable = Flowable.just(consentResult.getRight().get());
+            failReason = OutcomeReason.INTERNAL_ERROR;
         }
         else if(isOptedOut(consentResult.getLeft())){
-            logger.info("dpcMetric=OperationOutcomeReason,failReason={}", OutcomeReason.CONSENT_OPTED_OUT.name());
-            flowable = Flowable.just(buildOperationOutcome(patientID, IssueSeverity.WARNING,IssueType.SUPPRESSED,OutcomeReason.CONSENT_OPTED_OUT.detail));
+            failReason = OutcomeReason.CONSENT_OPTED_OUT;
+            flowable = Flowable.just(AggregationUtils.toOperationOutcome(OutcomeReason.CONSENT_OPTED_OUT, patientID));
         }else if(isLookBackExempt(job.getOrgID())){
             logger.info("Skipping lookBack for org: {}", job.getOrgID().toString());
 
             //TODO REMOVED THIS BEFORE MERGING!
             Random rand = new Random();
-            int randIndex = rand.nextInt(OutcomeReason.values().length-0) + 0;
-            OutcomeReason randReason = OutcomeReason.values()[randIndex];
-            logger.info("dpcMetric=OperationOutcomeReason,failReason={}",randReason.name());
+            int randIndex = rand.nextInt(OutcomeReason.values().length+1-0) + 0;
+            if(randIndex<OutcomeReason.values().length){
+                failReason = OutcomeReason.values()[randIndex];
+            }
             //END TODO
 
             flowable = Flowable.fromIterable(job.getResourceTypes())
@@ -89,9 +91,11 @@ public class JobBatchProcessor {
                 flowable = Flowable.fromIterable(job.getResourceTypes())
                         .flatMap(r -> fetchResource(job, patientID, r, job.getSince().orElse(null)));
             }else{
-                flowable = Flowable.just(LookBackAnalyzer.analyze(answers, patientID));
+                failReason = LookBackAnalyzer.analyze(answers);
+                flowable = Flowable.just(AggregationUtils.toOperationOutcome(failReason, patientID));
             }
         }
+        logger.info("dpcMetric=DataExportResult,dataRetrieved={},failReason={}",failReason==null,failReason==null ? "":failReason.name());
 
         final var results = writeResource(job, flowable)
                 .toList()
@@ -225,8 +229,7 @@ public class JobBatchProcessor {
             return Pair.of(consentService.getConsent(patientId),Optional.empty());
         }catch (Exception e){
             logger.error("Unable to retrieve consent from consent service.", e);
-            logger.info("dpcMetric=OperationOutcomeReason,failReason={}", OutcomeReason.INTERNAL_ERROR.name());
-            OperationOutcome operationOutcome = buildOperationOutcome(patientId,IssueSeverity.ERROR,IssueType.EXCEPTION,OutcomeReason.INTERNAL_ERROR.detail);
+            OperationOutcome operationOutcome = AggregationUtils.toOperationOutcome(OutcomeReason.INTERNAL_ERROR, patientId);
             return Pair.of(Optional.empty(), Optional.of(operationOutcome));
         }
     }
@@ -243,17 +246,6 @@ public class JobBatchProcessor {
             return optOutCount > 0;
         }
         return true;
-    }
-
-    private OperationOutcome buildOperationOutcome(String mbi, IssueSeverity severity, IssueType issueType, String text){
-        final var patientLocation = List.of(new StringType("Patient"), new StringType("id"), new StringType(mbi));
-        final var outcome = new OperationOutcome();
-        outcome.addIssue()
-                .setSeverity(severity)
-                .setCode(issueType)
-                .setDetails(new CodeableConcept().setText(text))
-                .setLocation(patientLocation);
-        return outcome;
     }
 
     private boolean passesLookBack(List<LookBackAnswer> answers){
